requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# misplmsi/research_mil/evaluators/deepset_eval.py: 146
# misplmsi/research_mil/evaluators/utils.py: 13
MulticoreTSNE == 0.1

# misplmsi/research_mil/evaluators/utils.py: 18,331,342
# misplmsi/research_mil/loaders/datasets.py: 8
# misplmsi/research_mil/loaders/utils_augmentation.py: 5
# misplmsi/research_mil/utils/post_seg.py: 14
# misplmsi/research_mil/wsi_tools/loader.py: 17
# misplmsi/research_mil/wsi_tools/sampled_spot_gen.py: 19
# misplmsi/research_mil/wsi_tools/tissue_mask.py: 9
# misplmsi/research_mil/wsi_tools/utils_augmentation.py: 5
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 19
Pillow == 8.3.1

# misplmsi/research_mil/test.py: 5
# misplmsi/research_mil/train.py: 7
# misplmsi/research_mil/wsi_tools/create_lib.py: 1
# misplmsi/research_mil/wsi_tools/sampled_spot_gen.py: 10
# misplmsi/research_mil/wsi_tools/tissue_mask.py: 6
# misplmsi/research_mil/wsi_tools/tumor_mask.py: 6
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 9
# misplmsi/research_mil/wsi_tools/xml2json.py: 2
PyYAML == 5.4.1

# misplmsi/research_mil/wsi_tools/tissue_mask.py: 13
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 26
histomicstk == 1.0.7

# misplmsi/research_mil/evaluators/deepset_eval.py: 144,145,148
# misplmsi/research_mil/evaluators/utils.py: 10,12,15
# misplmsi/research_mil/results_utils/collect_results.py: 12,13
# misplmsi/research_mil/wsi_tools/loader.py: 12
# misplmsi/research_mil/wsi_tools/sampled_spot_gen.py: 18
# misplmsi/research_mil/wsi_tools/tissue_mask.py: 12
# misplmsi/research_mil/wsi_tools/tumor_mask.py: 12
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 23,24
matplotlib == 3.2.1

# misplmsi/research_mil/results_utils/collect_results.py: 6
mlxtend == 0.17.2

# misplmsi/research_mil/evaluators/base_clf_test.py: 3
# misplmsi/research_mil/evaluators/base_eval.py: 8
# misplmsi/research_mil/evaluators/deepset_eval.py: 4
# misplmsi/research_mil/evaluators/utils.py: 3
# misplmsi/research_mil/loaders/datasets.py: 3
# misplmsi/research_mil/loaders/utils_augmentation.py: 2,3
# misplmsi/research_mil/results_utils/collect_results.py: 3
# misplmsi/research_mil/test.py: 6
# misplmsi/research_mil/train.py: 8
# misplmsi/research_mil/trainers/base_trainer.py: 8
# misplmsi/research_mil/utils/mil.py: 1
# misplmsi/research_mil/utils/misc.py: 3
# misplmsi/research_mil/utils/post_seg.py: 5
# misplmsi/research_mil/wsi_tools/annotation.py: 5
# misplmsi/research_mil/wsi_tools/loader.py: 9
# misplmsi/research_mil/wsi_tools/sampled_spot_gen.py: 17
# misplmsi/research_mil/wsi_tools/tissue_mask.py: 10
# misplmsi/research_mil/wsi_tools/tumor_mask.py: 8
# misplmsi/research_mil/wsi_tools/utils_augmentation.py: 2,3
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 12
numpy == 1.19.2

# misplmsi/research_mil/loaders/datasets.py: 2
# misplmsi/research_mil/wsi_tools/loader.py: 11
# misplmsi/research_mil/wsi_tools/sampled_spot_gen.py: 20
# misplmsi/research_mil/wsi_tools/tissue_mask.py: 5
# misplmsi/research_mil/wsi_tools/tumor_mask.py: 10
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 15
opencv_python == 4.3.0.36

# misplmsi/research_mil/wsi_tools/loader.py: 10
# misplmsi/research_mil/wsi_tools/sampled_spot_gen.py: 16
# misplmsi/research_mil/wsi_tools/tissue_mask.py: 11
# misplmsi/research_mil/wsi_tools/tumor_mask.py: 9
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 13
openslide_python == 1.1.1

# misplmsi/research_mil/results_utils/collect_results.py: 4
pandas == 1.1.5

# misplmsi/research_mil/loaders/utils_augmentation.py: 7,8
# misplmsi/research_mil/utils/post_seg.py: 12
# misplmsi/research_mil/wsi_tools/annotation.py: 6
# misplmsi/research_mil/wsi_tools/loader.py: 230
# misplmsi/research_mil/wsi_tools/utils_augmentation.py: 7,8
scikit_image == 0.17.2

# misplmsi/research_mil/evaluators/deepset_eval.py: 147
# misplmsi/research_mil/evaluators/utils.py: 14
# misplmsi/research_mil/results_utils/collect_results.py: 8
# misplmsi/research_mil/utils/post_seg.py: 9,10
scikit_learn == 0.23.0

# misplmsi/research_mil/results_utils/collect_results.py: 10,11
# misplmsi/research_mil/utils/post_seg.py: 4,8,13
scipy == 1.5.2

# misplmsi/research_mil/evaluators/deepset_eval.py: 149
# misplmsi/research_mil/evaluators/utils.py: 16
seaborn == 0.10.1

# misplmsi/research_mil/train.py: 14
tensorboardX == 2.2

# misplmsi/research_mil/evaluators/base_clf_test.py: 2
# misplmsi/research_mil/evaluators/base_eval.py: 2,3
# misplmsi/research_mil/evaluators/deepset_eval.py: 2,3,143,150
# misplmsi/research_mil/evaluators/utils.py: 2,19
# misplmsi/research_mil/loaders/__init__.py: 2
# misplmsi/research_mil/loaders/datasets.py: 4,6
# misplmsi/research_mil/models/__init__.py: 3,31
# misplmsi/research_mil/models/backbones/resnetv3.py: 1,2,3
# misplmsi/research_mil/models/pooling/classic.py: 1,2,3
# misplmsi/research_mil/models/pooling/deepset.py: 1,2,3
# misplmsi/research_mil/test.py: 6
# misplmsi/research_mil/train.py: 8
# misplmsi/research_mil/trainers/base_clf.py: 2,3
# misplmsi/research_mil/trainers/base_trainer.py: 2,3
# misplmsi/research_mil/trainers/deepset_trainer.py: 2,3
# misplmsi/research_mil/utils/misc.py: 2
# misplmsi/research_mil/wsi_tools/create_lib.py: 1
# misplmsi/research_mil/wsi_tools/loader.py: 15
# misplmsi/research_mil/wsi_tools/patch_gen.py: 6
# misplmsi/research_mil/wsi_tools/sampled_spot_gen.py: 12
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 18
torch == 1.9.0

# misplmsi/research_mil/evaluators/base_clf_test.py: 6
# misplmsi/research_mil/evaluators/deepset_eval.py: 7
# misplmsi/research_mil/evaluators/utils.py: 9
# misplmsi/research_mil/loaders/__init__.py: 3
# misplmsi/research_mil/trainers/base_clf.py: 6
# misplmsi/research_mil/trainers/deepset_trainer.py: 6
# misplmsi/research_mil/wsi_tools/loader.py: 14
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 17
torchvision == 0.10.0

# misplmsi/research_mil/evaluators/base_clf_test.py: 5
# misplmsi/research_mil/evaluators/deepset_eval.py: 6
# misplmsi/research_mil/evaluators/utils.py: 8
# misplmsi/research_mil/results_utils/collect_results.py: 2
# misplmsi/research_mil/trainers/base_clf.py: 5
# misplmsi/research_mil/trainers/deepset_trainer.py: 5
# misplmsi/research_mil/wsi_tools/loader.py: 8
# misplmsi/research_mil/wsi_tools/patch_gen.py: 7
# misplmsi/research_mil/wsi_tools/wsi_mil_seg.py: 11
tqdm == 4.62.2
